distributor: split HA dedupe middleware into shared scaffold + per-request strategy
Move prePushHaDedupeMiddleware from distributor.go to a new pkg/distributor/
ha_dedupe.go file, where it now only handles the work that all HA dedupe
strategies share: tenant resolution, AcceptHASamples gating, HA label lookup,
and active-group bookkeeping. The strategy is responsible for computing its
own sample timestamp (per-request uses a single request-wide value, per-sample
will need a per-replica minimum).

The strategy-specific body (span attributes, checkSample call, metrics
emission, __replica__ label stripping, next()) moves into the new file
pkg/distributor/ha_dedupe_per_request.go as d.perRequestDedupe(...), along
with the getEarliestSampleTimestamp helper that only the per-request path
calls.

No logic changes. The middleware wire-up position is unchanged.
