+++ pkg/distributor/ha_dedupe_per_request.go
@@ -1,0 +1,89 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package distributor
+
+import (
+	"context"
+	"strings"
+	"time"
+
+	"github.com/pkg/errors"
+	"github.com/prometheus/prometheus/model/timestamp"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
+
+	"github.com/grafana/mimir/pkg/mimirpb"
+)
+
+// perRequestDedupe decides whether to accept or dedupe an entire write request
+// based on the HA labels of the first series only. This is the legacy behavior
+// and assumes all series in a request share the same cluster/replica labels.
+func (d *Distributor) perRequestDedupe(ctx context.Context, pushReq *Request, req *mimirpb.WriteRequest, userID, haReplicaLabel, haClusterLabel, group string, now time.Time, next PushFunc) error {
+	cluster, replica := findHALabels(haReplicaLabel, haClusterLabel, req.Timeseries[0].Labels)
+	// Make a copy of these, since they may be retained as labels on our metrics, e.g. dedupedSamples.
+	cluster, replica = strings.Clone(cluster), strings.Clone(replica)
+
+	span := trace.SpanFromContext(ctx)
+	span.SetAttributes(
+		attribute.String("cluster", cluster),
+		attribute.String("replica", replica),
+	)
+
+	sampleTimestamp := timestamp.FromTime(now)
+	if d.limits.HATrackerUseSampleTimeForFailover(userID) {
+		sampleTimestamp = getEarliestSampleTimestamp(req, sampleTimestamp)
+	}
+
+	numSamples := 0
+	for _, ts := range req.Timeseries {
+		numSamples += len(ts.Samples) + len(ts.Histograms)
+	}
+
+	removeReplica, err := d.checkSample(ctx, userID, cluster, replica, sampleTimestamp)
+	if err != nil {
+		if errors.As(err, &replicasDidNotMatchError{}) {
+			// These samples have been deduped.
+			d.dedupedSamples.WithLabelValues(userID, cluster).Add(float64(numSamples))
+		}
+
+		if errors.As(err, &tooManyClustersError{}) {
+			d.discardedSamplesTooManyHaClusters.WithLabelValues(userID, group).Add(float64(numSamples))
+			d.costAttributionMgr.SampleTracker(userID).IncrementDiscardedSamples(req.Timeseries[0].Labels, float64(numSamples), reasonTooManyHAClusters, now)
+		}
+
+		return err
+	}
+
+	if removeReplica {
+		// If we found both the cluster and replica labels, we only want to include the cluster label when
+		// storing series in Mimir. If we kept the replica label we would end up with another series for the same
+		// series we're trying to dedupe when HA tracking moves over to a different replica.
+		for ix := range req.Timeseries {
+			req.Timeseries[ix].RemoveLabel(haReplicaLabel)
+		}
+	} else {
+		// If there wasn't an error but removeReplica is false that means we didn't find both HA labels.
+		d.nonHASamples.WithLabelValues(userID).Add(float64(numSamples))
+	}
+
+	return next(ctx, pushReq)
+}
+
+// getEarliestSampleTimestamp returns the earliest sample or histogram timestamp
+// across all timeseries in the request, or defaultTimestamp if no sample predates it.
+func getEarliestSampleTimestamp(req *mimirpb.WriteRequest, defaultTimestamp int64) int64 {
+	earliestSampleTimestamp := defaultTimestamp
+	for _, ts := range req.Timeseries {
+		if len(ts.Samples) > 0 {
+			if tsms := ts.Samples[0].TimestampMs; tsms < earliestSampleTimestamp {
+				earliestSampleTimestamp = tsms
+			}
+		}
+		if len(ts.Histograms) > 0 {
+			if tsms := ts.Histograms[0].Timestamp; tsms < earliestSampleTimestamp {
+				earliestSampleTimestamp = tsms
+			}
+		}
+	}
+	return earliestSampleTimestamp
+}
